+++ packages/client/src/devices/DeviceManager.ts
@@ -1,8 +1,9 @@
 import { combineLatest, Observable, pairwise } from 'rxjs';
 import { Call } from '../Call';
+import { TrackPublishOptions } from '../rtc';
 import { CallingState } from '../store';
 import { createSubscription } from '../store/rxUtils';
-import { InputMediaDeviceManagerState } from './InputMediaDeviceManagerState';
+import { DeviceManagerState } from './DeviceManagerState';
 import { isMobile } from '../helpers/compatibility';
 import { isReactNative } from '../helpers/platforms';
 import { Logger } from '../coordinator/connection/types';
@@ -20,8 +21,8 @@ import {
   MediaStreamFilterRegistrationResult,
 } from './filters';
 
-export abstract class InputMediaDeviceManager<
-  T extends InputMediaDeviceManagerState<C>,
+export abstract class DeviceManager<
+  S extends DeviceManagerState<C>,
   C = MediaTrackConstraints,
 > {
   /**
@@ -30,7 +31,7 @@ export abstract class InputMediaDeviceManager<
   stopOnLeave = true;
   logger: Logger;
 
-  state: T;
+  state: S;
 
   protected readonly call: Call;
   protected readonly trackType: TrackType;
@@ -43,7 +44,7 @@ export abstract class InputMediaDeviceManager<
     'filterRegistrationConcurrencyTag',
   );
 
-  protected constructor(call: Call, state: T, trackType: TrackType) {
+  protected constructor(call: Call, state: S, trackType: TrackType) {
     this.call = call;
     this.state = state;
     this.trackType = trackType;
@@ -109,7 +110,6 @@ export abstract class InputMediaDeviceManager<
 
   /**
    * Stops or pauses the stream based on state.disableMode
-   * @param {boolean} [forceStop=false] when true, stops the tracks regardless of the state.disableMode
    */
   async disable(options: { forceStop?: boolean }): Promise<void>;
   async disable(forceStop?: boolean): Promise<void>;
@@ -282,8 +282,11 @@ export abstract class InputMediaDeviceManager<
 
   protected abstract getStream(constraints: C): Promise<MediaStream>;
 
-  protected publishStream(stream: MediaStream): Promise<void> {
-    return this.call.publish(stream, this.trackType);
+  protected publishStream(
+    stream: MediaStream,
+    options?: TrackPublishOptions,
+  ): Promise<void> {
+    return this.call.publish(stream, this.trackType, options);
   }
 
   protected stopPublishStream(): Promise<void> {