+++ packages/react-native-sdk/src/utils/internal/rxSubjects.ts
@@ -1,5 +1,5 @@
 import { BehaviorSubject } from 'rxjs';
 
-export const isInPiPModeAndroid$ = new BehaviorSubject<boolean>(false);
+export const isInPiPMode$ = new BehaviorSubject<boolean>(false);
 
 export const disablePiPMode$ = new BehaviorSubject<boolean>(false);